Git issue 471: Use secondary thread-local caches
This changes the _cache back to holding compiled patterns,
rather than ._packed_data. This favors
single threaded operation, allowing the compiled patterns
to be used directly on the main thread with minimal
additional overhead for single-threaded operation. It is,
however, a bit less efficient for multi-threaded operation
as used patterns need to get the ._pickled_data once per
thread, rather than once per pattern.

For threads other than the main thread, each thread has
a secondary cache local to the thread.
If a compiled pattern is in the main _cache, then it
has to be copied into the thread local cache the first
time it's accessed, but is then served out of the
thread-local cache for any subsequent accesses.

Maintaining the thread-local caches' consistency
with patterns removed from the main thread's _cache
by _shrink_cache() is done by maintaining a count
of the number of times _shrink_cache() is called.
If it's been called more times than is recorded
for the current thread's thread-local cache, then
the patterns which are not in the _cache are
removed from the thread-local cache.